Add spell data to backend
Added various properties/classes for storing spells into the backend PathfinderSheet class. Now work needs to be done to add the data to the UI. Also fixed an issue if the player data has no photos. Additionally, changed the select panels to have a border, as it looked weird when they were all empty. I may mess with that later.