Finish all spell UI, improve file saving
Add UI for spell-like abilities. Fix some UI errors in Spell Editor. Added missing properties (spell conditional modifiers, spell specialty). All spell properties are now saved when the file is saved. Improved how sheets are saved: property names are now camelCase, and the order of JSON elements is now closer to how Mottokrosh's site does it (it is not exact, though). Made updates to UiCore to bring more stability/reliability to IntegerSpinner.

Current known issue: when interacting with the spell list, it moves the spell list to the top of the scroll viewer