Added email and photo list editors to UserDataEditor
Added editors for the email address list and photo URL list in the UserDataEditor dialog. With this, this dialog is now fully complete. Also added some missing functionality to the NewSheet dialog to make it mostly complete as well. Updated UiCore to include an improved SelectableItem control, which was used to allow editing the lists.