Update New Sheet layout, add banner image
Update the layout for the New Sheet layout to reduce spacing, so as to have it better match the visual style of this program. (The spacing in that dialog is more something that's fitting in Shine Calendar.) Also added a banner image to use for social media, and also fixed another crash. Changed "New Sheet" to "New Character" on File menu.