Set dirty if class skill checkbox is changed
Fixed bug where the sheet wasn't set as dirty if the skill's class skill checkbox was changed (checked or unchecked). Also attempted to fix issue where empty values would still be written in the generated JSON file, even if that particular skill was empty/undefined. However, all attempts to fix this issue resulted in either writing an empty value or writing "null" in the file. This seems impossible to fix due to the fact that I needed custom code to also support the "conditionalModifiers" element that Mottokrosh's website creates, and Json.NET can't seem to not write empty values when custom code is introduced. This is absolutely, completely annoying, and I'm upset there isn't a better way to handle this.