Add 96x96 icon
Based upon the 48x48 icon just doubled in size and cleaned up, this is the largest icon yet. The icon design itself, of course, comes from the design of a physical Pathfinder character sheet. (Upon further research, however, it seems that the 96x96 icon is actually not really useable for much in Windows. Whoops lol. I guess I'll have to upscale this icon at some point to make a 256x256 one.)