Added link to D20PFSRD for skills, update layout
Added an info button to the skill editors. Clicking on the info button will open up D20PFSRD for more info about that skill. Also made updates to the layout in the skill editor and on the General tab. Added tooltip to the Modifiers button in the skill editor (may remove later).