Added context menu for recent items
Added a context menu that will appear when you right-click on items in the Recent Items menu. This allows you to remove items individually or open the folder they are in. The implementation leaves a little bit to be desired, but it's pretty close to ideal. I also updated the "no sheet loaded" text to mention that the user can drop in a file. Also removed the "dotnet" file from the project itself, although it's still in the folder. I'm not quite sure what this file is here for, or how it got there...