Added a "wide state" setting for the skill editors
When the window is wide enough, the skill editors will enter a "wide state" mode, where all the modifiers will be visible at once without needing to open the Modifiers section. This will better utilize the space available when using the program with a wider window. (Tbh, a lot of the program doesn't exactly do much with the extra space if you make the window wider.)