Fix bugs with JSON view
If the JSON view's text is replaced or edited to no longer be valid JSON, the program will now silently delete and replace the JSON view's text with valid JSON from the sheet views. Also updated JSON view to update the dirty indicator if the text is changed. I worry that this may lead to some confusion if edits are made to the JSON view that make for invalid JSON, but I'd rather the program have the fail-safe of asking to save changes when they close the file.