Retarget project to .NET Core 3.1
PathfinderJSON now targets .NET Core 3.1 rather than just .NET Core 3.0. The newer version contains some bug fixes and improvements and also will be a LTS release, where as .NET Core 3.0 will be out-of-support in March. I suspect users will be sticking with the 1.0 release of PathfinderJSON for a little while, so the more stable I can make it, the better.