Rename "Sheet" menu to "Calculate", remove awaits
Hid the two refresh commands from the Sheet menu, as hopefully we won't be needing them anymore. Renamed "Sheet" to "Calculate" to make it more clear what the menu commands are for. Also added other comments in the code to better explain various things lol. Also removed awaits from various parts of the code as they were either not doing anything productive or not making a meaningful difference.