Added await to Update Calculations
When "Change Totals with Updates" is set, it notably slows down the calculation time... assumingly due to the multiple int.Parse functions wrapped in try-catch blocks going on. Wrapping them in an async/await Task doesn't make the process go much faster, but at least it allows the program to still be responsive during that time. This may open its own can of worms when attempting to do certain actions while the program is calculating, but that can be fixed in future commits. Another thing to add would be a visual indicator that the sheet is still calculating.