Prohibit certain actions while calculating
Prevent the app from being able to save or close the file while the sheet is calculating. While saving may not be too much of a stability concern, it may lead to the sheet being saved with only some calculations done and not others. Preventing the file from being closed (due to say directly closing the file, or closing it to open a new file) is for stability though, as it'd probably put the app in a weird state if the sheet was closed partway through calulcations. Other minor changes are present as well.