Remove irrelevant photo
Removed photo that was included in the repository from early on in the project's development. Not that it's a particularly revealing or bad photo, it shows a snapshot of me building this app early on, but it's not really an image that needs to be here. I moved it to a different location on my computer.