Add file drop support for JSON editor
Previously, users could not drop files into the program to open them, when it was in Raw JSON view. Now, that should be fine. Also, the JSON editor does support drag-and-dropping text into and out of it as well. (Try dragging in text from apps like Word, LibreOffice, or WordPad; Notepad doesn't support it.)