Color fixes in high contrast mode
When in Windows's high contrast mode and PathfinderJson's high contrast mode, there are some places where the colors clash or unexpected behaivor occurs making text hard to read. This fixes these issues to the extent that text is readable in most situations. There also appears to be an issue where text boxes in Windows's high contrast mode have a notably larger border when they're focused, but I'm not sure how to go about fixing that at the current time.