Add error handling for malformed Settings file
If the settings file is malformed or was somehow missing, PathfinderJson will create a new settings file. If PathfinderJson cannot create a new settings file (or create the initial directory), it will display an error and continue onwards. Will also display or fix errors if there are problems saving the settings file.