Add in keyboard shortcut code
This keyboard shortcut code will eventually be added into UiCore, but will be a part of this project for now while I'm building it out. These files come straight from Shine Calendar, but have been adjusted to be more general-use and to remove legacy Shine code. I also added in a MethodRegistry class, which can be used to keep a listing of the different event methods possible to be accessed via keyboard shortcuts; this allows me to simply iterate through all of the menu items and hook them up to keyboard shortcuts easily. Some additional work needs to be done, but a lot of the base stuff is there.