Finish keyboard shortcut support for PathfinderJson
Added a new helper class, KeyboardShortcutHandler, that directly sits between the window and the key registry. It can be easily added into any window to add keyboard shortcut support. Eventually, this will be all added into UiCore. Added MenuItemList class as well, but I'm not sure if I'll keep it around. Renamed "MethodRegistry" to "RoutedMethodRegistry" to give a more clear idea of what it contains. Added default keyboard shortcuts to PathfinderJson, but they can be overriden by an XML file in the AppData folder. I'll be writing a page on the Wiki to explain how to do this.