Added options icon, added more settings
Added a setting for word wrap in the text editor, as that was never kept before, and a setting to indent JSON in the text editor. Both are set to true by default, and will be reset to their default if you go back to 1.0.1, since that version isn't forward-compatible with these new settings. With the new text editor indent JSON setting, the existing "indentJsonData" setting will soon be used exclusively for when saving files. Also added a new options icon, and made a small tweak to how the Options window is displayed.