Add OK, Cancel buttons
Changed "color highlighting" to "syntax coloring". Trying to find the best wording to use that combines the standard term "syntax highlighting" with "color"... Anyway, also added an OK and Cancel button to the dialog. The OK button does actually save your settings, but they may not be reflected in the app right away.