Hook up Options dialog to make it work, removed "EditorIndent" setting
Completed hookups so when you click OK in Options dialog, it actually stores settings and then updates the main window. This is accomplished via a new ReloadSettings() function in MainWindow which allows it to reload the relevant settings any time after startup. Completed hookup so that the EditorWordWrap setting works. Removed EditorIndent setting as the save function is tied to the indenting of the editor, so these two settings cannot be currently separated. Thus, hid the Saving tab in Options as there's no unique option there anymore.