Update UiCore, improve CheckBox visuals in High Contrast mode
Updated UiCore, with improvements to how CheckBox appears while in High Contast mode. With this, PathfinderJson is now ready for release. The version of UiCore included here is technically not a finished version of 1.7, but the only other changes to 1.7 are not going to affect PathfinderJson in any way (as they involve controls that PathfinderJson does not use). After that's done, I hope to release UiCore!