Fix issue with dropping in files when no file is loaded
Files can now be dropped into the raw JSON view when there isn't currently a file open. This bug came as the result of two intersecting things: the raw JSON editor only accepts files when it is enabled, and it is disabled when a file is not open. Because of this, that prevented the raw JSON editor from being able to accept dropped files in this state.