Re-arrange options window
Re-arranged layout of items in options window (nothing was moved from one tab to another). Expanded height of options window to allow more space to work with, and also to fit the new "Show line numbers" option in the JSON Editor options. Re-built the text preview. Also increased the default height in the main window to fit everything in the General tab without needing to scroll, when the toolbar is shown.