Created HTML template file for later use
This HTML file will be the template for when I implement the Export to HTML feature later. The file is mostly unfinished, but a lot of the framework is there, and I'll be adding to it some more until it's all complete. When the feature is complete, PathfinderJSON will dynamically generate the final HTML file using this template and the data from the actual sheet itself. Note: this feature is not planned until version 1.2, so don't expect it until then. The next version planned is still 1.1.2.