Update IntelliCode model
Re-trained IntelliCode model to match 1.1.2. This hasn't been re-trained in a while, and I honestly don't know if I'll be able to continue re-training it for every release. For now, I think I'll still aim to do so, but don't rely upon it. There was a change in my development setup and now Visual Studio doesn't always give me access to the IntelliCode model page.