Rework spacing in General/Skills/Items tabs
Reworks spacing in the General, Skills, and Items tabs to better account for situations like the "Make text bigger" Ease of Access option. Other tabs and editors will come along shortly. There will be some graphical glitches and weird things to figure out along the way with this as well, but the minimum desire is that text isn't cut off with larger text sizes.

Also, the previous commit did downgrade MarkDig back to 0.18.0, to be compatible with MarkDig.Wpf. I'll probably make my own fork of MarkDig.Wpf to get it working with the new MarkDig releases (I at least want to get the performance improvements of MarkDig 0.20.0).