Added custom search panel
This custom search panel replaces what comes with AvalonEdit. The colors and appearance match the rest of PathfinderJson's interface now, and it also is slightly more accessible than before. Unfortunately, due to how AvalonEdit set up its searching functionality, some of the code had to be duplicated from AvalonEdit and put into this repo. I may in the future try a pull request where some of these classes are opened up so I don't have to replace them. Also, Regex searching doesn't seem to work too well? I'll need to investigate if this issue comes from something in the process of me duplicating the code from AvalonEdit, or if it is an issue core to how AvalonEdit wrote it. This also included an updated version of Ssui, which has improvements to FlatButton and SelectPanel.