Re-design layout/spacing for New Sheet window
Re-did the layout and spacing of the New Sheet window, to match the other windows and controls done thus far. For some reason, the window, when I set it to dynamically set its own height, wants to be waaaay longer than it actually needs to be. I wanted it dynamically set, but that won't work, so I've needed to add a scroll bar (which shouldn't display by default). Also, I unified the appearance of the modifier text boxes across the entire program now.