Clear player's profile image between opening sheets
If you open one sheet that contains a player's profile image, and then you close it and open a second sheet that does not contain the player's profile image, the program would still display the previous player's image. This has been fixed so that if the player for the current sheet doesn't have a set profile image, no image is shown.