Update HTML template
Change the preferred fonts for the template, add in the feat table display, make some more adjustments for how the sheet looks when printed. I will be making some layout and design changes later on as I'm not quite happy with how this is turning out. Also note that the HTML Export and printing feature will not make it for 1.2; instead, I'll be making a later update that includes these features.