Lock Options window's font size to 12
For the sake of releasing version 1.2 sooner, it's easier to simply lock this window's font size to 12 rather than attempting to re-design the window to make its spacing flexible like the rest of the program. This, unfortunately, means the Options window won't work with the "Make text larger" accessibility option and there may also be some weird layout issues when Windows's High Contrast mode is enabled. In a later version, I plan to do a complete re-design of the Options window as a result