Changed Notes Markdown setting to a sheet setting
The Notes Markdown option is changed over to a sheet setting, rather than being stored separately in the JSON. Being stored as a sheet setting makes the toggle backwards compatible with 1.1.1 and 1.1.2 (not 1.1.0, sadly). If you enabled Markdown in some of your files using some of the earlier builds, it will go back to being disabled with this build onwards. Just re-check the box again and it'll be fine! In the future, I'll also be adding in a toggle to always use Markdown in Notes.