Added "File Saved" text that displays when... the file is saved
This text appears for 15 seconds, but you can click on it to make it go away instantly. This displays when you save the file yourself or if an auto-save occurs. This confirms to the user that there was an auto-save, rather than them having to blindly trust the program's auto-save feature is on.