Improve how player data is loaded after being edited in main window
After editing the user data (player data), the main window tries to load the first email even when one doesn't exist. This issue is now fixed. The main window also attempts to load in the first profile photo if one was added, or removes the profile photo if all of them were removed.