Add "Keep on Top" function to dice roller
Add a "Keep on Top" button to the dice roller window. When enabled, the dice roller window will not be hidden by other windows when you focus on them. Also, made it so that only one dice roller window can be created (per PathfinderJson instance). If you select the Dice Roller command while a window is already open, it brings the existing window back to the front and in focus.