Create flexible UndoStack class
This is my own personal UndoStack class that I built, that I'll use here. It uses a generic type, so any item can be stored and used for undoing/redoing. It is, indeed, a state-based undo system. Also created a barebones window to test out the feature, and after a few bug fixes, it seems to be working good. Next step is getting it set up with the MainWindow itself!