Added Edit menu to sheet view, undo/redo not working yet
There are now two Edit menus: one that is shown in the sheet views, and one that is shown in the text editor view. This is due to the fact that the text editor maintains its own undo/redo system, that is separate from the undo/redo system I developed here. Also, the text editor commands are not relevant in sheet view.

In addition, while debugging, there is an option to open the Undo Stack testing window on the sheet Edit menu. In some future commit, I may remove this command and window.