Replace SkillEditor with the new SkillEditor
The new SkillEditor was designed in SentinelsJson and was intended to replace what's currently in PathfinderJson. So that's happening now. The new editor is a lot more flexible, not only in regards to being easily adaptable to different TTRPG systems, but also in allowing the users to change which modifier is used for a skill. Ideally, this new editor should be easier to read and understand as well. There's still some work I need to do (such as handling the modifier change), but 90% of it is now done! At some point, PathfinderJson will be switched over to use Skill Lists too, like SentinelsJson, rather than the hardcoded system currently in SkillEditorFactory.