Setup Undo/Redo
The Undo/Redo function now works! ... Except that it creates an undo state for every single individual character you type in a text box. It gets a bit ridiculous. I tried to fix it via the undoSetTimer, but it's not working quite right. So I'll have it in this ridiculous state right now and then I'll fix it up in a future commit.

The Undo/Redo buttons will enable themselves when you're able to Undo or Redo.