Added Skill Lists to PathfinderJson
Added skill list support to PathfinderJson, similar to how it is in SentinelsJson. The biggest difference is that PathfinderJson utilizes a special SkillEditorFactory class; for right now, I decided to simply modify the SkillEditorFactory to use skill lists rather than recreate the entire skill system used in SentinelsJson.

The SkillEditorFactory will by default load the skill list included in PathfinderJson with the standard skills, but if a sheet has a specific skill list file listed, the SkillEditorFactory can find it and use it instead.

With this, custom skill lists can be used just like in SentinelsJson. This opens the door for customization where otherwise not possible (such as with third-party classes that add extra skills). A proper Skill List editor will come to both PathfinderJson and SentinelsJson soon.