Read custom skill modifiers from files
When the user sets a custom modifier for a skill (via the "Use stat" drop-down for that skill), PathfinderJson will now properly re-apply it when you re-open the file.

The changes were already being properly saved to a file, and I had written the code to read these changes from a file when it's loaded... but I forgot to connect that code to the loading method. Whoops!