Changed spell table/money to numeric spinners
The text boxes in the table of the top of the Spells tab, and in the money section of the Items tab have been replaced with IntegerSpinners and DoubleSpinners respectively.

As I mentioned on Twitter, PathfinderJson 1.3 is moving away from exact compatibility with Mottokrosh's website. If the text boxes are used as intended, compatibility should still be maintained, but if you entered text into these specific boxes, they will be replaced with zeroes.

PathfinderJson 1.2.1 will remain in order to maintain exact compatibilty with Mottokrosh's site. I may also make a quick bug-fix release as well to port over some fixes currently in 1.3.