Fix differences between skill editors
The skill lists feature expects the usage of the new Skill Editor present in version 1.3, so I had to modify the files so they work properly with the old skill list.

As of right now, PathfinderJson 1.2.2 will now support skill lists! However, this is relatively basic support; if you define a custom skill list file, PathfinderJson 1.2.2 will load it, otherwise it'll load the default list every time. The custom file can only be set via JSON (in the sheetSettings element). No visual editor or GUI for skill list support, as these are all things coming with 1.3.

In essence, this is a backport of the 1.3 feature for the sake of future compatibility, but for those who want to get down and dirty, it's useful here now too.