Create Counters feature
Counters will be one of the big features of the 1.3 release. In this commit, they are now totally functional and nearly all complete. I may make some adjustments later on to add support for changing the background color of the circle and fix some other visual things, but yeah.

The idea of a counter is to assist with items where you need to count up and down something, such as an effect that lasts 5 turns or even tracking a character's health.

Also, fun little feature: while holding down the Ctrl key, the counter will go up and down by 5 rather than 1.