Visual improvements for Counters
Made some visual improvements for how counters look. Also added a border onto the Counters window, to visually separate the counters from the toolbar above them.

In a future commit, I'll add support for importing and exporting counters and the option to have PathfinderJson automatically store your counters for when you exit and re-open the program.