Expand edit mode functions of CounterItem
When you enter edit mode now, you can change not only the counter title, but also the color and you can directly set the value as well!

I may in the future streamline this even further, but for here and now, it'll work. One fix I do need to make is updating the icons for edit mode: currently Save and Undo are used (for "apply" and "cancel" respectively), but I want to instead replace them with a Check icon and either Cancel or Back.