Add file import/export support to Counters
Now all counters can be exported to a JSON file to be loaded later (like, for example, saving them for in between sessions).

At this point, the counters feature is nearly entirely complete. I could technically call it complete now, but I'm debating adding in one more feature or not. We'll see! (I may also change the default background color for new counters.)