Added Add Setting dialog to Sheet Settings
Now users will be able to insert new sheet settings without needing to dive into the JSON editor. The dialog is structured to also allow me to use it as an Edit Setting dialog, but I'm not doing that right now.

Note that there isn't any error handling for if an existing setting already exists with this name, but that's okay. When the settings list is saved, whatever is the lowest (latest) one on the list will be the one actually saved, as it'll overwrite previous values.