Update to .NET 6.0, Solid Shine UI 1.9
This moves PathfinderJson from the old ESR .NET Core release up to the latest ESR .NET release.

I didn't move to .NET 5.0 in the past due to concerns about SSUI not working with it and concerns about the PublishSingleFile option not working as I wanted in Windows. I've corrected SSUI's issues, and I think the .NET team corrected the issues with PublishSingleFile.

Also updated to SSUI 1.9, including the new SelectPanel and FileSelect controls. I've fixed compiler errors that occurred due to the breaking changes with the new SSUI release. In a future commit, I'll move away from the obsolete SelectPanel methods.