Fix new nullability issues that occurred in .NET 6.0
More things were marked for nullability with .NET 6.0, which surfaced more code points where null values could occur that I hadn't accounted for. Now they are accounted for! ... For the most part.

I'll need to take a peek at AvalonEdit's code to fix the nullability issue with the search panel.