Improved layout, functionality of SheetSettings dialog
Decreased the empty space on the left side of the Sheet Settings dialog, and made sure everything lined up correctly. Also added a confirmation dialog before deleting a value.

Also added some convenience features and a small bit of data validation to the Add Setting dialog, such as pressing Enter to accept and close the dialog (or Escape to cancel). Also changed the layout a bit here too, to remove some awkward spacing.