Change SkillEditor Misc property to string
To better match the old skill editor and Mottokrosh's editor, I changed the Misc property to a string, and put a textbox for it in the editor.

The text box will grow in width a bit while text is put in, but it will then afterwards grow in height, allowing a user to put in as much text as they would like.

Also, if the user enters in just a number in the text box, it will be counted towards the calculation.