Fix Undo-related bugs, decrease undo delay timer
Fix a few bugs related to Undo and Redo not correctly retaining/reloading data, and not clearing the stack when it should be cleared.

Also decreased the undo delay timer by 1 second; this creates more undo states overall, but this also means less time is spent before an undo state is created. The delay before was too long where you type in some text, and you have to wait for the timer before you can Undo it.