Add "icon view" for the base ability scores
Added in an icon view, where each score is displayed in a little flag with the value editors and stuff in/around them. Stands out and looks a bit more flashy than the standard table from before. You can switch back to table view if desired, though!

Also added a link to D20PFSRD about ability scores, and also fixed an issue with adding in temporary ability score values not marking a sheet as dirty.