Add temporary editors for ability score icon view
When you click on the Temp button, the temporary editors will be displayed below all of the icons. This was the best way I could do this without getting even wilder with the layout than I already did. It's not the best UI design, in my eyes, but it'll have to do.

Also added in code to make sure the temporary values stick when you switch between ability score views! Whew!

Finally, also fixed a visual bug I introduced last time, whoops!