Add info section about Armor Class being calculated
Added a small area that is shown when the "Update Armor Class with Updates" calculate option is turned on. With this option on, the Armor and Shield values in the Armor Class editor get updated based upon the values of the armor and shields below. This area explains that option and what happens.

This is in response to issues that I've personally had where I didn't understand why the Armor Class is being auto-calculated.

Also I've realized that calculation settings aren't stored in any way, and instead default values are loaded in every time you open the program. I may have to change that...