Create ColorSchemeDialog to change color scheme
This will replace the various functions currently in place to allow users to change the color scheme. Note that the dialog isn't fully complete, and there's other further work to be done to fully integrate it into the UI. Also note that some changes will not be preserved when the program is closed and re-opened. But at least now a bulk of the visual work is done, now it's just a lot of backend stuff.

Also fixed a few issues in relation to the ItemEditor not supporting the Quantity value as a string.