ColorSchemeDialog options all properly set
Now, if you select a light theme or dark theme, or a high contrast theme, from the Color Scheme Dialog, the setting is now set properly and remembered. And it is also backwards-compatible with previous versions, as earlier versions will default to a colored theme instead.

Merged the "Light theme" and "Dark theme" options, into a new "Basic theme" section. This should also help simplify the dialog a decent amount. Also resized the dialog to make it a bit shorter.

Also removed the "High Contrast" option in the View menu, as the new dialog will set the theme properly.