Store calculation settings in sheet settings
When a calculation option from the Calculate menu is changed, a setting is added to the current sheet to retain that change. That way, users can store calculation settings with each sheet that they make, and change them as needed.

In the future, I'll add an option to set default calculation settings.

Also fixed sheet settings not applying when the Sheet Settings dialog is opened or saved. Finally, made some optimizations for loading sheet settings.