Saving sheet settings dialog now marks sheet dirty
Now when you click Save in the Sheet Settings dialog, the sheet is marked as dirty (has unsaved changes), prompting the user to save if needed.

There are currently no checks to actually make sure if the newly saved sheet settings actually differs from the previous settings, but this at least errs on the side of saving/preserving data rather than discarding it (unlike earlier commits).