Add Psionics skil list as option in Sheet Settings
I didn't realize how unfinished I left the Sheet Settings dialog before lol. Now the dialog properly attempts to load in the skillList setting, and if the "psionics" list is put in, then there is a Psionics Unleashed option that is selected.

Also now when saving the Sheet Settings dialog, the entire skill list is stored and then reloaded (in case the "skillList" property has changed).