Revert SkillEditor to older (1.2.2) version
The new skill editor doesn't have the full backwards compatilbity that is expected of the 1.2.x releases, so I've had to replace it back with the older skill editor.

I've made a few adjustments so that in theory, the old skill editor should be able to load in skills with unconventional modifiers without issue, but I'll have to test that in a bit.

Also put a max height on the Misc text box, because that was starting to get ridiculous.