Rewrite skill modifer change via Sheet Settings
When the "skillModSet" setting is modified in the Sheet Settings dialog, now the skill list is properly updated and set with the new list of modifiers.

However, I've just realized this currently breaks reloading the skill list in case the "skillList" setting is changed. So I'll need to work on this a bit more.