Rearrange Interface Options screen
Replaced the various Color Scheme and High Contrast Mode options with the button to open the Color Scheme Dialog (similar to what was already done on the Main Window).

Also added on a checkbox to show/hide glyphs in various editors in the main window, although there's currently no functionality tied to it yet.