Change CompoundEditor to use SheetValueEditor
Changed CompountEditor to use the new DetailsButton and the SheetValueEditor for its values. This makes the calculations code also decently simpler. As a result, I dropped the calculation code being done async in the CompoundEditor as it doesn't seem worth it anymore.

Also changed ArmorClassEditor to use the SheetValueEditor as well.