Change spell data table to use SheetValueEditor
Now, rather than IntegerSpinners all over this, SheetValueEditors are used instead, which allows users to enter in non-numerical text if desired.

By default, some of the values will be filled with 0s if there isn't a different value put in; however, the 0s are not saved to the file (to match behavior of previous versions).

Also some minor layout changes and a change to a recent files error dialog.