Add SettingsIo file, proper portable release support
This represents a major overhaul over how PathfinderJson handles user settings. This will open up two major things:

Firstly, a truly portable version is now possible, by storing the settings next to the .exe file. This can be transferred between computers or handled in whatever way is deemed necessary. A flag in SettingsIo determines if it's a portable release or not, and SettingsIo handles that as needed. Note that some data is still stored on whatever computer runs this program, but it's nothing major that needs transferral between computers.

Secondly, by storing the settings file in separate directories for each version, it's now possible to have multiple versions running and store separate settings for each version (granted, only for each version after this commit). This will be used in the future when I set up an installer for people to easily install PathfinderJson on their system, rather than downloading zip archives like before.

I think doing all of this stuff at the start of the program ended up messing up the WPF system, so I moved all of this logic out of the App constructor and into the Application_Startup event. We'll see how that affects the ProfileOptimization, I suppose. However, I honestly am not doing that much more than I was before; the biggest change is detection of old app settings.