Set many dialogs to appear in center of parent window
Most (if not all) dialogs appear at a random location on the primary monitor, rather than appearing on top of (or at least near) the parent window that spawned the dialog.

With this, now many of these dialogs (I don't think I've reached all of them) will appear directly on top of their parent window.

For people with only 1 monitor, this probably doesn't make much of a difference. For multiple monitor setups though, this can help fix a weird annoyance.