Added option to hide side tabs in Continuous Scroll view
When in Continuous Scroll view, you can hide the tabs on the side if you want. That way, you can focus more on the sheet itself. However, this option is only available in this view (as it's obviously needed in tabbed view, and is unusable in Raw JSON view).

I actually envisioned adding this option years ago, but for some reason, just never did it. Well, here we are!