Refactor SettingsIo, move all File IO to SettingsIo
Use SettingsIo from now on to load and save settings, as this will handle locating and saving in the settings directory for us. Added some more methods and XML documentation to help describe the code, as well as other refactoring to make it more suitable to bring to other apps.

Now the overhaul of the Settings handling is complete!