Adding IEquatable to various base types in PathfinderSheet
With IEquatable, now it's easier for me to compare two items, and overall make it easier to compare two PathfinderSheet objects. While this probably doesn't have that much usage for just editing values in PathfinderSheet, I can use this as a part of debugging or improving my code (such as getting the Undo/Redo functionality all set up).