Setup filtering for ItemListDisplay
I haven't actually tested it yet, so this could be completely broken as far as I know. But at least the code and structure is all there!

To make this work, I've applied the filter to the IldPropertyInfo object, since this is generated per each ILD, and each ILD only maintains one copy of all the IldPropertyInfo objects, so this can be consistently referred back to. From there, I just created the code to match the filter.

The hardest part was putting together the final parts of UI to enable creating the other types of filters. As a part of this, I added on a NumberInputDialog - which uses 1-2 DoubleSpinners to use for applying values.

From here, I'll do some testing soon, and then see where else I could potentially improve this. But at least this is the last major feature done!