Beginning work of rewriting SettingsIo
I wrote this file a few years ago before I actually had a solid idea of how Qhuill was going to function, so a few things have changed in between these times. I think I was also just heavily focused on getting *something* done, so that I could say it's done (and potentially use as a reference for the future)... without putting as much thought as I should about whether this was the best built interpretation of what I wanted to do.

There is also right now the issue that this code keeps offering to import old settings even though there are now existing current settings. In a future commit, I'll be more fully rewriting  the LocateLatestSettingsDirectory and SetupSettingsBase methods to fix that bug and also just have code that makes more sense on the surface.

In this commit, I reworked the constant values at the top and also reworked the process for creating and establishing the directories that are used by PathfinderJson.

This one code file might become so monolithic that it might be better to separate out into multiple files, we'll have to see.