SSUI 1.9.8, use FlatRepeatButton in CounterItem
Updated to use SSUI 1.9.8. Various updates to CounterItem: I'm using the FlatRepeatButton for the plus and minus buttons, which allows users to hold down the button to slowly count up, rather than needing to repeatedly click.

I also reduced the corner radius for the edit buttons at the bottom, since it felt a bit out of style for what my usual thing is. Finally, I added a colors image next to the Change Color button to better indicate what that button is for.