Beginning work to debug/work more on Undo/Redo
Implemented a number of things so I can start doing some debugging of the undo/redo feature in the future. Still a bit more work to do with this, but yeah.

I'm really wanting to figure out how exactly the Undo/Redo system is working, and where I'm tripping over something.