Improvements to undo/redo functionality
I finished implementing the debugging features for viewing the undo stack, and now I can see what the stack is at any time. With that in mind, I was able to make some adjustments.

Now, a new undo state is created when a file is opened or created, so that we can continue undoing back to the original state. That was one of the core missing elements, I've found out.

Also, I've implemented IEquatable for SpellLevel, which helps prevent it from appearing as a changed item in the undo stack. The entire Skills dictionary also keeps appearing as being changed, even when it hasn't, which is frustrating, but I won't tackle that at this point.

Now, the undo/redo feature feels a bit more solid, and I think I'm ready to roll with it. There's still an odd quirk where it'll create an undo state when you first type into a box, as well as when you stop typing into it, as well as any change up or down you make to a spinner. In the future, I might want to tackle why that occurs, but I'll save that for later.